Renderer - Avoid binding textures that a shader doesn't use
The render pass code was previously binding diffuse and normal map
textures, even when the shader was not using them.

Add a check prior to binding textures so that textures are only bound
when the shader features require them.

This change takes performance test average frame time from 10.79ms to
10.77ms